many more changes - most notably prototypes and structs so that we are now dealing with floats

+++ structs.h
@@ -2,7 +2,7 @@
 #define STRUCTS_H
 
 typedef struct vertex {
-    int x, y, z, w;
+    float x, y, z, w;
 } vertex ;
 
 struct linseg {
